requirements.txt
absl-py==2.4.0
anyio==4.12.1
astunparse==1.6.3
audiomentations==0.43.1
audioread==3.1.0
backports.tarfile==1.2.0
build==1.4.0
CacheControl==0.14.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cleo==2.1.0
cmsisdsp==1.10.2
contourpy==1.3.3
crashtest==0.4.1
cycler==0.12.1
decorator==5.2.1
distlib==0.4.0
dulwich==0.25.2
fastjsonschema==2.21.2
filelock==3.20.3
findpython==0.7.1
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.13.2
keyring==25.7.0
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.46.0
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
murenn==0.3.0
namex==0.1.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
numpy-minmax==0.5.0
numpy-rms==0.6.0
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pandas==3.0.0
pbs-installer==2026.1.27
pillow==12.1.0
pkginfo==1.12.1.2
platformdirs==4.5.1
poetry==2.3.1
poetry-core==2.3.0
pooch==1.9.0
protobuf==6.33.5
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-stretch==0.3.1
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
TextGrid==1.6.1
threadpoolctl==3.6.0
tomlkit==0.14.0
torch==2.10.0
trove-classifiers==2026.1.14.14
typing_extensions==4.15.0
urllib3==2.6.3
virtualenv==20.36.1
Werkzeug==3.1.5
wrapt==2.1.0
xattr==1.3.0
zipp==3.23.0
zstandard==0.25.0
